UIQM-348: Edit a MARC authority record | Do not allow user to delete 010 field
