+++ src/QuickMarcEditor/QuickMarcEditor.js
@@ -48,7 +48,9 @@ import {
   getCorrespondingMarcTag,
   getContentSubfieldValue,
   deleteRecordByIndex,
-  are010Or1xxUpdated,
+  is1XXUpdated,
+  is010Updated,
+  is010$aPopulatesBibField$0,
 } from './utils';
 import { useAuthorityLinking } from '../hooks';
 
@@ -159,10 +161,21 @@ const QuickMarcEditor = ({
       return;
     }
 
-    if (marcType === MARC_TYPES.AUTHORITY && linksCount > 0 && are010Or1xxUpdated(initialValues.records, records)) {
-      setIsUpdate0101xxfieldsAuthRecModalOpen(true);
+    if (marcType === MARC_TYPES.AUTHORITY && linksCount) {
+      if (is1XXUpdated(initialValues.records, records)) {
+        setIsUpdate0101xxfieldsAuthRecModalOpen(true);
 
-      return;
+        return;
+      }
+
+      if (
+        is010Updated(initialValues.records, records) &&
+        is010$aPopulatesBibField$0(initialValues.records, instance.naturalId)
+      ) {
+        setIsUpdate0101xxfieldsAuthRecModalOpen(true);
+
+        return;
+      }
     }
 
     handleSubmit(e)
@@ -179,6 +192,7 @@ const QuickMarcEditor = ({
     getState,
     validate,
     showCallout,
+    instance,
   ]);
 
   const paneFooter = useMemo(() => {
@@ -417,6 +431,7 @@ const QuickMarcEditor = ({
                     subtype={subtype}
                     marcType={marcType}
                     linksCount={linksCount}
+                    instance={instance}
                   />
                 </Col>
               </Row>
@@ -489,7 +504,7 @@ QuickMarcEditor.propTypes = {
     reset: PropTypes.func.isRequired,
   }),
   marcType: PropTypes.oneOf(Object.values(MARC_TYPES)).isRequired,
-  linksCount: PropTypes.number.isRequired,
+  linksCount: PropTypes.number,
   locations: PropTypes.arrayOf(PropTypes.object).isRequired,
   httpError: PropTypes.shape({
     code: PropTypes.string,
