+++ src/QuickMarcEditor/QuickMarcEditorRows/utils.test.js
@@ -7,48 +7,6 @@ import { MARC_TYPES } from '../../common/constants';
 import * as utils from './utils';
 
 describe('QuickMarcEditorRows utils', () => {
-  describe('isReadOnly', () => {
-    it('should be true for record start', () => {
-      expect(utils.isReadOnly({ tag: '001' })).toBeTruthy();
-    });
-
-    describe('when record type equals BIB', () => {
-      it('should be true for tag 004', () => {
-        expect(utils.isReadOnly({ tag: '004' })).toBeFalsy();
-      });
-    });
-
-    describe('when record type equals HOLDINGS', () => {
-      it('should be true for tag 004', () => {
-        expect(utils.isReadOnly(
-          { tag: '004' },
-          QUICK_MARC_ACTIONS.EDIT,
-          MARC_TYPES.HOLDINGS,
-        )).toBeTruthy();
-      });
-    });
-
-    it('should be true for tag 005 (Date and Time of Latest Transaction)', () => {
-      expect(utils.isReadOnly({ tag: '005' })).toBeTruthy();
-    });
-
-    it('should be true for record end', () => {
-      expect(utils.isReadOnly({ tag: '999', indicators: ['f', 'f'] })).toBeTruthy();
-    });
-
-    it('should be false for record rows', () => {
-      expect(utils.isReadOnly({ tag: LEADER_TAG })).toBeFalsy();
-      expect(utils.isReadOnly({ tag: '002' })).toBeFalsy();
-      expect(utils.isReadOnly({ tag: '050' })).toBeFalsy();
-      expect(utils.isReadOnly({ tag: '998' })).toBeFalsy();
-      expect(utils.isReadOnly({ tag: '999' })).toBeFalsy();
-    });
-
-    it('should be true for tag LDR on derive page', () => {
-      expect(utils.isReadOnly({ tag: '005' }, QUICK_MARC_ACTIONS.DERIVE)).toBeTruthy();
-    });
-  });
-
   describe('hasIndicatorException', () => {
     it('should be true for exeptional row', () => {
       expect(utils.hasIndicatorException({ tag: '001' })).toBeTruthy();
@@ -76,42 +34,6 @@ describe('QuickMarcEditorRows utils', () => {
     });
   });
 
-  describe('hasDeleteException', () => {
-    it('should be true for exeptional row', () => {
-      expect(utils.hasDeleteException({ tag: LEADER_TAG })).toBeTruthy();
-      expect(utils.hasDeleteException({ tag: '001' })).toBeTruthy();
-    });
-
-    it('should be true for record end', () => {
-      expect(utils.hasDeleteException({ tag: '999', indicators: ['f', 'f'] })).toBeTruthy();
-    });
-
-    it('should be false for exeptional row', () => {
-      expect(utils.hasDeleteException({ tag: '010' })).toBeFalsy();
-    });
-
-    describe('when record type equals BIB', () => {
-      it('should be true for tag 245', () => {
-        expect(utils.hasDeleteException({ tag: '245' }, MARC_TYPES.BIB)).toBeTruthy();
-      });
-    });
-
-    describe('when record type equals HOLDINGS', () => {
-      it('should be true for tag 004', () => {
-        expect(utils.hasDeleteException({ tag: '004' }, MARC_TYPES.HOLDINGS)).toBeTruthy();
-      });
-    });
-
-    describe('when record type equals AUTHORITY', () => {
-      it('should be true for 1XX tags', () => {
-        expect(utils.hasDeleteException({ tag: '150' }, MARC_TYPES.AUTHORITY)).toBeTruthy();
-      });
-      it('should be true for tag 010 that populates bib field(s)', () => {
-        expect(utils.hasDeleteException({ tag: '010' }, MARC_TYPES.AUTHORITY, 1)).toBeTruthy();
-      });
-    });
-  });
-
   describe('hasMoveException', () => {
     it('should be true for exeptional row', () => {
       expect(utils.hasMoveException({ tag: LEADER_TAG })).toBeTruthy();
@@ -124,36 +46,6 @@ describe('QuickMarcEditorRows utils', () => {
 
     it('should be false for common rows', () => {
       expect(utils.hasMoveException({ tag: '010' }, { tag: '011' })).toBeFalsy();
-    });
-  });
-
-  describe('isMaterialCharsRecord', () => {
-    it('should be true for exeptional row', () => {
-      expect(utils.isMaterialCharsRecord({ tag: '006' })).toBeTruthy();
-    });
-
-    it('should be false for common rows', () => {
-      expect(utils.isMaterialCharsRecord({ tag: '010' })).toBeFalsy();
-    });
-  });
-
-  describe('isPhysDescriptionRecord', () => {
-    it('should be true for exeptional row', () => {
-      expect(utils.isPhysDescriptionRecord({ tag: '007' })).toBeTruthy();
-    });
-
-    it('should be false for common rows', () => {
-      expect(utils.isPhysDescriptionRecord({ tag: '010' })).toBeFalsy();
-    });
-  });
-
-  describe('isFixedFieldRow', () => {
-    it('should be true for exeptional row', () => {
-      expect(utils.isFixedFieldRow({ tag: '008' })).toBeTruthy();
-    });
-
-    it('should be false for common rows', () => {
-      expect(utils.isFixedFieldRow({ tag: '010' })).toBeFalsy();
     });
   });
 });
