+++ src/QuickMarcEditor/utils.test.js
@@ -1755,4 +1755,87 @@ describe('QuickMarcEditor utils', () => {
       expect(utils.are010Or1xxUpdated(initial, updated)).toBeTruthy();
     });
   });
+
+  describe('hasDeleteException', () => {
+    it('should be true for exeptional row', () => {
+      expect(utils.hasDeleteException({ tag: LEADER_TAG })).toBeTruthy();
+      expect(utils.hasDeleteException({ tag: '001' })).toBeTruthy();
+    });
+
+    it('should be true for record end', () => {
+      expect(utils.hasDeleteException({ tag: '999', indicators: ['f', 'f'] })).toBeTruthy();
+    });
+
+    it('should be false for exeptional row', () => {
+      expect(utils.hasDeleteException({ tag: '010' })).toBeFalsy();
+    });
+
+    describe('when record type equals BIB', () => {
+      it('should be true for tag 245', () => {
+        expect(utils.hasDeleteException({ tag: '245' }, MARC_TYPES.BIB)).toBeTruthy();
+      });
+    });
+
+    describe('when record type equals HOLDINGS', () => {
+      it('should be true for tag 004', () => {
+        expect(utils.hasDeleteException({ tag: '004' }, MARC_TYPES.HOLDINGS)).toBeTruthy();
+      });
+    });
+
+    describe('when record type equals AUTHORITY', () => {
+      it('should be true for 1XX tags', () => {
+        expect(utils.hasDeleteException({ tag: '150' }, MARC_TYPES.AUTHORITY)).toBeTruthy();
+      });
+
+      it('should be true for tag 010 that populates bib field(s)', () => {
+        const authority = { naturalId: 'n79018119' };
+        const initialValues = { records: [{ tag: '010', content: '$a n  79018119' }] };
+        const linksCount = 1;
+
+        expect(utils.hasDeleteException({ tag: '010' }, MARC_TYPES.AUTHORITY, linksCount, authority, initialValues)).toBeTruthy();
+      });
+    });
+  });
+
+  describe('isReadOnly', () => {
+    it('should be true for record start', () => {
+      expect(utils.isReadOnly({ tag: '001' })).toBeTruthy();
+    });
+
+    describe('when record type equals BIB', () => {
+      it('should be true for tag 004', () => {
+        expect(utils.isReadOnly({ tag: '004' })).toBeFalsy();
+      });
+    });
+
+    describe('when record type equals HOLDINGS', () => {
+      it('should be true for tag 004', () => {
+        expect(utils.isReadOnly(
+          { tag: '004' },
+          QUICK_MARC_ACTIONS.EDIT,
+          MARC_TYPES.HOLDINGS,
+        )).toBeTruthy();
+      });
+    });
+
+    it('should be true for tag 005 (Date and Time of Latest Transaction)', () => {
+      expect(utils.isReadOnly({ tag: '005' })).toBeTruthy();
+    });
+
+    it('should be true for record end', () => {
+      expect(utils.isReadOnly({ tag: '999', indicators: ['f', 'f'] })).toBeTruthy();
+    });
+
+    it('should be false for record rows', () => {
+      expect(utils.isReadOnly({ tag: LEADER_TAG })).toBeFalsy();
+      expect(utils.isReadOnly({ tag: '002' })).toBeFalsy();
+      expect(utils.isReadOnly({ tag: '050' })).toBeFalsy();
+      expect(utils.isReadOnly({ tag: '998' })).toBeFalsy();
+      expect(utils.isReadOnly({ tag: '999' })).toBeFalsy();
+    });
+
+    it('should be true for tag LDR on derive page', () => {
+      expect(utils.isReadOnly({ tag: '005' }, QUICK_MARC_ACTIONS.DERIVE)).toBeTruthy();
+    });
+  });
 });
